Fix task 1 test errors

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -10,15 +10,15 @@
 void print_numbers(const char *separator, const unsigned int n, ...)
 {
 	unsigned int count;
-	unsigned int hold;
+	unsigned long int hold;
 	va_list arg;
 
 	va_start(arg, n);
 
 	for (count = 0; count < n; count++)
 	{
-		hold = va_arg(arg, unsigned int);
-		printf("%d", hold);
+		hold = va_arg(arg, unsigned long int);
+		printf("%ld", hold);
 		if (separator != NULL)
 			printf("%s", separator);
 	}
